add utility ToGua file for matrix & quat conversion
